[testsharder] Don't cd into the build directory
This effectively removes global state from testsharder, as code no
longer assumes that the current directory is the build directory. This
makes testsharder easier to understand and test.

I ran testsharder locally with the `-image-deps` and `-hermetic-deps`
flags against a core.x64-asan build directory and confirmed that the
output is the same before and after this change.

Change-Id: I0742e7bf79974265af7f90121ff7f177f0551236
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691144
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>