[boot-shim][depthcharge] Specially configure the UART
This is effectively a reland of fxrev.dev/683850 for everything not
loaded by depthcharge. It introduces a hack for the depthcharge case,
which will soon be removed in favour of a proper CMDLINE-bases fix
described in the associated bug.

Bug: 101395, 102726
Change-Id: If8eddd4ba0ca1ed24d458791813d74484419b4a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688826
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>