[ot-radio] Changes to fix ot-radio TX delay
There is an issue in the current handling of interrupt which results in
hold-off of ~40msec in packet to be transmitted. This manifests as a
test failure in bug 231208021, but affects most SPI transactions.
It is found that TX was held off due to RX taking long time because of
the delays in each interrupt. Modified delays to make faster exit from
interrupt handling.

Currently, removing delays for first 5 do-while attempts in receivng
transaction. This takes care of the common case where 2 transactions are
needed for receiving long message - first transaction confirms the size
of inbound packet, and second transaction is done with the size. This
was taking ~10msec at ToT because of the delay in do-while loop.

Also added an early exit for spurious interrupt event. Since the inner
loop can take care of multiple interrupt events (while handling the
interrupt the pin gets deasserted and reasserted if packet is still
present on the other end of SPI). This results in an interrupt event
getting queued up.

Bug: b/231208021

Tests: 1) ot-radio driver test.
       2) Thread Harness joiner and Leader test cases.

Change-Id: I7b63b87678d3e9c748a278478cf0a7c99416ef8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690346
Reviewed-by: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Parag Dixit <paragdixit@google.com>