[services] Fix flake of filtered_service_routing_test
Accesing a non-existent or filtered out service instance can fail with a NOT_FOUND error when creating the
protocol proxy (opening the protocol member on the service proxy) or
when calling the method on the protocol proxy. This depends on when the
channel used to open the service instance directory is closed by the
underlying service provider.

Multiply: fuchsia-pkg://fuchsia.com/component-manager-services-tests#meta/filtered-service-routing-test.cm
Fixed: 101997
Change-Id: Ifda9e8fb9f240686d74af3f4dc1a75c256feb63e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691347
Commit-Queue: Ryan Diaz <ryandiaz@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>