[minfs] introduce minfs component
This introduces the minfs component. Similar to the blobfs component, it
starts a fuchsia.fs.startup.Startup protocol when the component is
started, which is used to further parameterize it with the block device,
options, and operation to perform. The lifecycle protocol is also
implemented.

This adds two new sets of tests, one which tests the ComponentRunner
implementation directly, and the other which actually launched the
component with the cml as a dynamic child and calls the startup protocol
on that.

Bug: 101655
Test: fx test minfs-component-runner-test minfs-component-test
Change-Id: Ib8468605400677bd0aa8a71d2a72bdce15f27cd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686914
Reviewed-by: James Sullivan <jfsulliv@google.com>
API-Review: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>