[scripts/owner] Upload OWNERS files for review
Extend `suggest_owners.py --generate_missing` to commit the generated
OWNERS files and upload the change for review. Add as reviewers all the
newly added owners.

Add a `--dry_run` argument to print the content of the generated OWNERS
file and the git commands without creating the file or uploading the
change for review.

Ask for confirmation before uploading for review if it's not a dry run.

Drive-by: write to cvs file only if `--cvs` argument exists.

Change-Id: I1246aa4445508a6d6cd0d685165b91ef56ded6aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691402
Commit-Queue: Marina Ciocea <marinaciocea@google.com>
Reviewed-by: Chris Palmer <palmer@google.com>
Reviewed-by: Shai Barack <shayba@google.com>