[vulkan][compact-image] Fix decommit on trim
If there are multiple imports of a buffer, each of them most be unpinned
before the decommit is legal. This happened to work before because
within a single driver multiple imports were being deduplicated, so an
unpin of one unpinned all of them. That doesn't happen anymore.

Change-Id: Ie9a3a56c5366b27bfba560274aba6628c23165e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691748
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: John Bauman <jbauman@google.com>