[network/tests] Introduce multi-nic test environment
This introduces a multi-homed network environment that tests may run
on to validate socket operations when a stack has multiple
interfaces/routes to a destination.

A client environment is created with 2 interfaces (on the same broadcast
domain) and a server which will simply respond to any received UDP
packets with the same data but with "Response: " prepended to it.

Change-Id: I524b6e247373423dfec547779ed2a916b14fc0b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691306
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>