[docs][hub] Remove 'deleting' dir documentation
The Hub's `deleting` directory was removed in Change-Id
I58da854b73981838799004f1c89002e435619c0f.

This CL removes references to the `deleting` directory from Hub
documentation. There is no longer a "staging" directory for a node that
is to-be-deleted in Hub.

The reason for this CL is update documentation from changes made to Hub.

Bug: 93481
Change-Id: I711c3f1a050b32269911addf9bfe63c648cd8a0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691843
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Dave Smith <smithdave@google.com>