[socket] Add functions for reacquiring signals
There's a Rust socket optimization where READ/WRITE signals stay
set until a read or write fails.

Some applications require that polling for socket readable or socket
writable actually means that the socket is readable or writable before
trying to act on it, so an API is needed that clears and reacquires
the signal lets us poll and know that the signal is "fresh".

Example use case for virtio-vsock:
1) Guest tracks socket buffer and never sends more data than the socket
   can handle, so a write never fails and the signal is never cleared
2) The device wants to wait on socket writable to inform the guest that
   there is room with a credit update
3) Polling always returns Ready since (per 1) the signal is never
   cleared
4) We tight loop spam the Guest with credit updates telling it that
   there's no credit

The solution is to have the ability to clear the signal before polling,
so we know that the polled signal is current. This won't be used
generally as most applications can just let the write fail.

Change-Id: Ibf61315eeebbc62f4eb2c4e2a5a33d5525bf0ffe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691004
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>