[testsharder] Add max shards per env test case
It's important that testsharder try not to exceed `-max-shards-per-env`,
so add a test case that covers the path where it needs to recalculate
the target shard duration to keep the shard count low.

`-max-shards-per-env` is technically not a hard maximum and we will
slightly exceed it for environments that are split up by affectedness
and hermeticity, which is why there are three non-skipped shards in the
resulting golden file even though `-max-shards-per-env` is 2.

Also add a `testListEntry()` helper function to make it more concise to
mark tests as hermetic in the test data.

Change-Id: Id795d7c71ba3ef1f68fc4126a74b6fb29c25b7b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691271
Reviewed-by: Ina Huh <ihuh@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>