[netstack] Add trace points at netstack boundaries
Introduce trace points at the boundary between the netstack and client
applications (through the `fuchsia.posix.socket` API), the boundary
between the netstack and gVisor, and the boundary between the netstack
and drivers (in the ethernet and netdevice clients).

For now, only async events are emitted. NB: it's necessary to use async
events (as opposed to regular trace durations) because the Go runtime is
free to move goroutines between the backing OS threads, and asynchronous
tracing is necessary to correlate the start/end of an event that spans
multiple threads.

Also, no flow tracing is used to correlate the flow of a particular
packet through the system; this might be added in the future.

Test: Follow instructions in src/connectivity/network/TRACING.md to
      enable tracing in the network realm, and run `ffx trace start`.
      Verify that the resulting trace includes events in the `net`
      category.
Fixed: 99538

Change-Id: I2a87bfd825d893fe115ad8e363be7521a09eaa2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688426
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>