[slimfast] Memory budget boot test
Introduce a host-target test that boots Fuchsia to a session in a
RAM-limited VM.

In the course of this work, discovered that fshost doesn't seem to be
able to mount a virtual disk pointed at `storage-full` image (a raw FVM
block image) because it runs out of space. Turns out our dev tools
already handle this by creating a temporary upsized image when the tool
runs:
https://cs.opensource.google/fuchsia/fuchsia/+/main:tools/devshell/qemu;l=141-147

This patch also adds a mechanism for tests to do the same.

Bug: 100672

Change-Id: Ia35c583bfcb49364160d6f6adba8721ab04775d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683825
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>