[a11y] Re-enable web semantics tests for hit testing and scrolling
These tests assume that the root semantic node specifies the scale
factor used to convert between the client view's allocated and logical
coordinat spaces. However, that scale factor is not available to the
client until it receives a view metrics event from scenic, so I believe
it's possible that the root node may not yet contain the scale factor
when it's first committed to fuchsia.

This change gates our web spatial semantics tests on receipt of the
correct scale factor to avoid that race condition.

Bug: 81626, 99748
Test: semantics-integration-tests
Multiply: semantics-integration-tests: 500

Change-Id: I10b87289d0c13039abdda8dd1b14e88031c19e71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691205
Commit-Queue: Alexander Brusher <abrusher@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>