[ot-radio] Fix inadvertent interrupt signalling
Found that InvokeInterruptHandler was called in normal operation. This
was supposed to be called only when we stopped handling interrupts
because allowance went to 0. On a subsequent increment of allowance,
we wanted to trigger Interrupt Handler. But it was getting called
inadvertently when we set the flag to true in the inner do-while loop,
and in next iteration of do-while loop, another thread increases
allowances, find the flag to be true and Invokes interrupt handler.

Also made the boolean atomic since it's shared between the two threads,
and writing to it only if there is a change.

Bug:  b/231208021

Tests: 1) ot-radio driver tests
       2) Thread Harness joiner and Leader test cases.

Change-Id: I5e66eac218afecd48c615945d8d38d2164b21ae3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691525
Commit-Queue: Parag Dixit <paragdixit@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>