[tefmocheck] Allow for truncated except block.
The end string of an except block may appear after the last test
finishes and accidentally get truncated from the logs since we don't
have a way to know to wait for it. In that case, we should assume that
if we see the start string of an except block in the logs but no end
string after, it was probably truncated and so everything from the start
string to the end of the log should be included in the except block.
Except block strings should be chosen carefully so that you wouldn't
normally see them in the logs except as part of an except block.

Bug: 102747
Change-Id: I089283cb6cf78d04e36886b7df2bca9b7cdf0516
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692168
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>