[netstack] Benchmark socket API with fake netstack
Add a variant of the socket benchmarks that runs against a "fake
netstack", which does the minimum possible amount of work to implement
the subset of the `fuchsia.posix.socket/Provider` protocol needed for
the benchmarks.

By comparing with the same benchmarks for the production netstack, this
will give us a signal about how much of the latency in e.g. UDP is due
to work done in the core netstack vs. due to the overhead of the API
(IPC, kernel roundtrips, FIDL de/serialization, etc.).

The fake netstack currently implements support for UDP and ICMP echo
sockets. Support for TCP and potentially "Fast UDP" will be implemented
in future changes.

Fixed: 98501
Test: fx test socket-benchmarks-with-fake-netstack
Change-Id: Ia10a798aab623261a4b21185fcd5222124313e47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670947
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>