[blobfs] Change blob loader so that it no longer attaches VMOs
Calling AttachVmo currently has the potential to trigger a deadlock with
the driver stack if something else is calling GetTopologicalPath on the
same driver host that contains Blobfs's underlying block device.

To avoid this, change BlobLoader so that it only attaches at mount time.

An unfortunate side effect of this is that it will involve copying
merkle tree data. However, this is more than likely offset by the fact
that we don't have to wait for an AttachVmo round-trip when loading
blobs.

This change may have a positive impact on memory since memory-resident
merkle trees won't have to be padded to the system page size.

Fixed: 102743
Change-Id: Ia9c8f023277666ec2be682eeaa6d26ae345ff7f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692482
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Chris Suter <csuter@google.com>