[iwlwifi] Ported TKIP algorithm
Ported the algorithm from the standard. The integration code
are coming in the later CL.

Note that this CL only contains the algorithm to generate P1K
(Phase 1 Key). P2K is implemented in the firmware.

Bug: 92939
Tested: added unit test from the test vector on the standard.

Change-Id: Ifdd306b605cc228db89bd7ca4178ae854a21b192
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667142
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Louis Lo <yjlou@google.com>