[intel-i915] Rename display_count -> display_config_count; use std::span.
The name "display_count" used in intel-i915.cc is misleading: usually
it stands for the count of display configs passed in by
ApplyConfiguration instead of the actual number of active displays /
pipes, which could cause error when we calculate the DBUF size
(will address in a subsequent CL).

This CL just renames all the "display_count" to "display_config_count"
to reflect this fact, and replaced a lot of (raw pointer, count) usages
with std::span, which is more clear in semantics.

Bug: 95863
Test: Variable renaming + minor refactoring. No-op.
Change-Id: I51b09f5fb73cbbd3ba2b1c8ac8fcdd0bcbcb49e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691530
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>