[ffx-emu]Adding engine_type property to runtime config
This exposes the engine_type as part of the runtime config object
so that it can be used when special-casing flags based on the engine.

A concrete example of this is touchpad input is currently only supported
on aemu, but the goal is to add support later so this should not be
hard-coded into the launcher.

Bug: 99762

Change-Id: I22dfaaddeab4592ede0d4dea42d897575ec8cbed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681563
Reviewed-by: Matt Hamrick <mjhamrick@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>