[kernel] Change OOM timeout defaults
Increase OOM timeout default to 50 seconds.

For the bringup product set the timeout to 8 seconds. Bringup is
a small product and we'd expect reboots to be quick, if it takes
longer than 8 seconds, something is very likely wrong.

For the core product use the default.

Userspace now shuts down in an orderly fashion and is expected to
restart the device promptly so the kernel timeout functions as a
backstop.

Empirically all product configurations almost always reboot in
less than 10 seconds so we believe if the system actually takes
greater than 50 seconds something is truly wrong in userspace.

Fixed: 91744
Change-Id: I258fd7bd774a47c68a3255c7cb00705b5573ffef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692170
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>