[audio][services][mixer] Fix error on server shutdown during tests
During tests, we expect that all server shutdowns should be initiated by
the server (by closing the channel and sending an epitaph) or by the
client (by closing the channel). Any other shutdown logs an ERROR:
https://cs.opensource.google/fuchsia/fuchsia/+/main:src/media/audio/mixer_service/fidl/base_fidl_server.h;drc=f35f9ae3acd9998458627174f8416b3b54b83a5b;l=96

For example, if the dispatch loop is stopped before a server is
shutdown, this is logged as an ERROR. To avoid this error, tests wait
for servers to shutdown before stopping the dispatch loop. However, our
tests currently have a bug we wait for top-level servers
(FidlSyntheticClockRealm) but NOT for child servers which are created by
those top-level servers (FidlSyntheticClock).

To fix this bug, BaseFidlServer is augmented with a notion of "child"
servers and BaseFidlServer::WaitForShutdown is updated to wait for all
children to shutdown.

This CL also adds a helper class, TestServerAndClient, to simplify this
test logic. TestServerAndClient owns a local server and client
connection. When destroyed, the class destroys the client connection
then waits for the server to shutdown.

Fixed: 101589
Change-Id: I374a70c686c6f4c4e0002be238627d95efd9789f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692164
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>