[virtualization] Implement WipeData in LinuxManager
Implement WipeData by writing zeroes to the first 1 MiB of the guest
stateful FVM partition. This is enough to corrupt the filesystem data-
structures which causes linux to reformat the device.

I initially implemented this by Deleting the FVM volume, however this
doesn't actually work in practice because when the FVM is subsequently
recreated, it can be allocated the same disk slices.

Bug: 92508
Test: fx test linux_runner_tests
Change-Id: I7745ffc50a52a8a10868dc373488c06efa2e7960
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691117
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Rupert Shuttleworth <ruperts@google.com>