[fshost] launch minfs as a component
Make fshost launch minfs as a static child component. This mostly
piggybacks on the fxfs fshost changes, so it's mostly a cml change, and
changing filesystem-mounter to call the component launching logic.

This also adds a change to the minfs component runner to skip some steps
on shutdown if we don't have a valid pointer to a minfs, so we don't try
to call things on a minfs instance that isn't configured yet.

Fixed: 101655
Bug: 91577
Change-Id: Ib29bbc5626e8579130601532f0410523b82ee78d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686949
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>