[onet] Remove onet.cmx
The tool isn't very useful, and there's no evidence of recent use.

Also remove onet from fuchsia_shell_package. Keep the tool target for
use in a test (it may be ok to delete the test, but we don't have to to
delete the cmx).

Change-Id: I743006569bfa731ab06bd80e4b42cec73aa50930
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691572
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>