[element_manager] Allow using Realm from parent
This CL changes the semantics of the `element_manager`
component to use the `fuchsia.component.Realm` protocol
from parent instead of from framework. This is done in order
to enable product variability of the capabilities routed to
the elements collection. Furthermore, the name of the collection
itself is now configurable using structured configuration.

In order to aid a soft transition, a component with the old
semantics and name will be kept until downstream users are moved
to using the new component.

Bug: 95841
Bug: 100933
Change-Id: I7ce6da279058475b8afc84b4929a3cba2eaf4670
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691451
Reviewed-by: John Shamoon <johnshamoon@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>