[realm_builder][rust] allow local components to get stop notification
Adds a feature to rust realm builder with which local components can
register a stop notifier. Once this task is performed, any stop requests
that realm builder receives for the local component is reported over the
stop notifier, and it is then the local component's job to exit.

As with any component, if the component does not exit in a timely
fashion component manager will request the component be killed, which
causes realm builder to stop the local component immediately.

Fixed: 82021
Change-Id: I93d3e6ab68df9c9c83735e58b17aac79a5f116d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692683
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>