[ermine] Migrate to new element_manager
This CL changes the element_manager component used
for launching apps to the newer one. Semantically,
the difference is that element_manager consumes
`fuchsia.component.Realm` from its parent component,
as opposed to the framework. This is done to allow
product variability in the capabilities routed to the
elements launched.

This CL is the first of two. This one changes the component
URL for element_manager to `...new_element_manager.cm` along
with the aforementioned changes. Once the older version of
element_manager is removed, the URL will be changed back
to `...element_manager.cm`.

Bug: 95841
Change-Id: Ifbd5935fbe65edd896f81bcd07503580171b212b
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/691391
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>