[vaapi] Fix fuzzer error
If buffer allocation fails during a mid stream buffer reallocation then
the method CoreCodecMidStreamOutputBufferReConfigFinish() will never be
called and instead CoreCodecStopStream() will be called. Ensure that we
stop waiting if either is called. Also if output_re_config_required is
false when onCoreCodecMidStreamOutputConstraintsChange() is called then
the buffers are not reallocated, meaning that we should not wait since
CoreCodecMidStreamOutputBufferReConfigFinish() will not be called and
instead the new buffer constraints will be sent instead.

Bug: 103142
Change-Id: I260b5a2a9d99eb292534085c73970c3775db85f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693807
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>