[diagnostics][persistence] v2 integration test
Port of the v1 Persistence integration test to v2. As part of this,

In Persistence:
 - After publishing the persisted data, Persistence adds a property
    "published" at inspect root with a value of monotonic now().
    This helps avoid race conditions from partially published data.
 - Now uses Structured Config

In SamplerTestController:
 - Make IncrementInt synchronous
 - Add SetOptional and RemoveOptional to support more test cases

In iquery:
 - Delete src/diagnostics/iquery/test/test_component/meta/test_component.cmx

PS1 of this CL is the same as fxr/691602 which flaked and was reverted.
PS2 of this CL fixes the race condition.

Multiply: fuchsia-pkg://fuchsia.com/persistence-integration-tests#meta/persistence-integration-test.cm
Fixed: 84283
Fixed: 102672
Change-Id: Ifc16aa694df2ec5432ad577c3a4bc8ac93144004
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693488
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Chris Phoenix <cphoenix@google.com>