[audio_core] Route audio protocols directly from CFv2 audio_core
This is step 5 of the sequence described at fxbug.dev/90408#c20.

This CL also moves soundplayer out of src/sys/core/meta/core.cml since
soundplayer is not included in all products (such as recovery images).
This change is necessary in this CL because soundplayer depends on
audio_core, which is not in all products and therefore is not in
core.cml.

Bug: 90408
Change-Id: Icaa1daf22ef03c51f7325630003578e3debe1e10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665949
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>