[build][assembly] Only allow the fuchsia toolchian in //build/images
Stumbled across this by accident.  This change adds an assert that the
//build/images/BUILD.gn file is only processed in the Fuchsia toolchain,
not the host (or any other variant), as well as correcting the one place
that was causing it to be processed in the host toolchain.

Change-Id: I98a1d06f5f8a2c9c879728bb11613a9bbc86e33d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687270
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>