[netstack3] Swap NS2/NS3 in images via GN flag
Add a flag to control which of two Fuchsia packages is used as the
system's network component. `fx set --args=use_netstack3=true ...`
builds with Netstack3 in place of Netstack2 as the system netstack. The
default behavior is to use Netstack2.

In addition, add missing services that the system netstack is required
to implement to Netstack3's manifest (even though Netstack3 does not
serve these yet) and a new Netstack3 component with the same name and
manifest URL as the Netstack2 component

Bug: 100736
Change-Id: Ibc8cf0c539865db7f2e35ec852c7eb8345abb51d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686314
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>