[vaapi][fuzz] Fix error in fuzzer
When I added fuzzing to the tiled output of the VAAPI decoders I did not
change the buffer size calculation. Therefore it was possbile for the
codec to given buffers that were not big enough to save the requested
image data. We must have two different pic_size calculations: one for
linear output and one for tiled output.

Bug: 103141
Change-Id: Iabe1ed1f286d911a72661d95c4d6316e50d5b6d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693816
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>