[assembly] Establish contract for base-resolver config verifier
This change implements a simple golden file check for the new
base-resolver structured configuration. It includes a custom
implementation to deal with the need to have separate goldens for
"eng" and "non-eng" build types, and vacuously integrates the
implementation into assembly. The integration is vacuous because
it always compares a golden file to itself with TODOs in the
assembly build files to add the correct file paths when the
base-resolver structured configuration lands in fxrev.dev/687102.

Bug: 102968
Change-Id: I4bb1ef25209e9057782f45ae8643270e598737e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694083
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Fuchsia-Auto-Submit: Mark Dittmer <markdittmer@google.com>