[kernel] Hide version string from compiler
kVersionString is filled in outside of the source file that defines
this variable. The compiler during lto picked up that because it is
const the size should always be 3, but this isn't actually the case.
By just hiding any provenance information about the return value of
version_string from the compiler we force it to actually call
strlen at runtime with the real value.

Fixed: 103245
Change-Id: I40291d02f29bb091c4316737b7d6b6c07efc6824
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693800
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Alex Brachet <abrachet@google.com>
Fuchsia-Auto-Submit: Alex Brachet <abrachet@google.com>