[scenic][screen_capture] get buffer_count from importer
This CL updates the ScreenCaptureBufferCollectionImporter to map the
collection_id to the buffer count that will allow screen capture to
access the buffer_count instead of having the client import it with the args.

This CL also updated the renderer type from before where in would take
in a VkRenderer and instead now takes in any rederer type that inherits
Renderer. This allows for more flexibility.

Test: Tests ran on physical device and new tests added
$ fx test screen_capture_unittests

Change-Id: I32fe4dd5f98984345172846fc177c5e693c2a440
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691790
Commit-Queue: Oliver Lance <ollylance@google.com>
Reviewed-by: Felipe Archondo <felipearchondo@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>