[audio][clock] Move Timer into SyntheticClockRealm
The mixer service has a Timer class which will be used to wait for the
next scheduled mix job. There are two implementations, one based on real
clocks and one based on syntehtic clocks. This CL refactors the
synthetic clock implementation to be driven by a SyntheticClockRealm.

Bug: 87651
Change-Id: I578f8274e41d0fd4d0856b8468aa1d1601375314
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691578
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>