[netstack3] Add UDP multicast subscription in Core
Add the ability to register for multicast membership via a socket
handle. The device used to register is chosen based on arguments passed
in and the device the socket is bound to, if any. If no device is
selected, the implementation will fail to register - the Linux-like
behavior of picking a device via the routing table will be implemented
in a future commit.

Bug: 39479
Change-Id: I0413d42ded51298555b36312e97bace4d63de894
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/690147
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>