[kernel][vm] Plumb write flag on the VMO pin path
Pass down the IOMMU write permission to the VMO pinning code so that it
can distinguish between pinning for a read and a write. This is
essential to dirty track pages in the case of a pinned write. Once the
pages have been pinned and mapped in the IOMMU, the VM system will no
longer be able to trap the writes in order to mark pages dirty.

This commit just plumbs through the write flag but does not actually use
it yet. That will follow in a future CL. There is no change in behavior
with this commit.

Bug: 63989
Run-All-Tests: true
Change-Id: Ia18b14c3b15d93f991c193234ae8bb62d07c82a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693382
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>