[fdf][ddk] Add driver_runtime_shared_lib sdk item
Change the driver_runtime implementation's shared_library to an sdk
shared library so that it can be consumed out of tree. This is needed
because in order to unit test drivers out of tree we need to link the
test binaries against symbols that are present in this shared library.

Bug: 102050
Tests: No behavior changes so no tests are needed
Change-Id: I6e786e3758b488936cf28f286db0e9ac5cccb9e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694443
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>