[kernel][coresight] Do not constrain enumerated components
Previously we would error out if the components found in a ROM table
were not of "CoreSight class" or that of a particular legacy timestamp
generator. In this change, we remove that assertion, as such logic is an
overfit and does not leave room for leniency in encountering other
legacy components or vendor-specific wonkiness. A ROM table walker
should just be in the business of reporting its entries; validation of
expected components is more reasonably deferred to users.

Bug: 103310
Change-Id: I9f6d5aebfc9eaaa8ba6b9472443539413ec66622
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694686
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>