[revert][as370] Undelete drivers related to the Synaptics AS370 SOC
This is a mechanical revert of 21d77ff05881f8af8a5c6e0f87608abd587e91cd.

These drivers will not build in the current form but are not connected
to the build system. Will apply diffs on top of this base CL to restore
functionality.

Bug: 94099

Change-Id: Ia04f315b300acb062595e9150acde765d4ea86c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693934
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>