[virtualization] Fix race in vmm start
Move host_vsock back to the guest_manager to fix the race between
GetHostVSockEndpoint and LaunchGuest.
Both calls could start VMM and depending on timing VMM might end up
calling GuestConfigProvider::Get before config is initialised.
LaunchGuest is a only way which is supposed to start VMM, but
GetHostVSockEndpoint would have start host_vsock which would try to
connect to fuchsia::virtualization::GuestVsockEndpoint exposed by VMM
and start VMM.
If GetHostVSockEndpoint win the race, VMM would use the default initialised config and subsequentely
trigger assert crash because guest_memory was not set.
Issue had a higher repro chance right after pave, but it was possible to repro
it after OTA as well.
Host vsock endpoint used to be inside the guest manager and it was moved to a standalone component for CFv2 migration and this CL is just moving it back into the CFv2 guest managers.

Test:
- Manually pave / ota workstation and launch LinuxTerminal.
- Add a basic unit test to cover hostvsock endpoint - guestvsock endpoint interaction

Bug: 103242
Change-Id: I5740f7f3d1885ead930104375d9cd5d37f95396b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694730
Commit-Queue: Alexander Radchenko <alerad@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Adam Barth <abarth@google.com>