[virtualization] Migrate in-process vsock tests
Migrate the existing in-process test suite to support the in development
out-of-process device.

Note that the new device doesn't support the full functionality of the
old device, so many test cases are behind an #if. Most of these won't
build, as they require updated helper functions and boilerplate.

Bug: 97355
Change-Id: Ib9ed1050c0073cee964dd338c682f5a63b27de80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686153
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>