[settings] Add ManagedInspectMap utility
The existing InspectMap can be included in structs that implement
Inspect and adds all of its children to the inspect tree once when
attached. Our existing use cases have tended to insert and remove items
dynamically, which ends up requiring work to add the children to the
inspect hierarchy anyways. See the attached bug for more context.

This wrapper around HashMap automatically writes inserted items to the
inspect node that it manages. The existing InspectMap is renamed to
InspectWritableMap to better reflect its functionality, as items
inserted into it have to be manually written to inspect.

One usage of InspectWritableMap is migrated to ManagedInspectMap, along
with some test additions and cleanup. Most of the other usages of
InspectWritableMap can be migrated to ManagedInspectMap or simply
replaced with plain HashMaps.

Bug: 103390
Test: fx test setui_service_tests
Change-Id: I399833af58cf3e1240b20a8facab41a478006556
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694515
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Commit-Queue: William Xiao <wxyz@google.com>